将double换为bigdecimal，方便前端处理

+++ src/main/java/org/zhuzhu_charging_station_backend/dto/ChargingStationResponse.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import org.zhuzhu_charging_station_backend.entity.ChargingStationSlot;
 import org.zhuzhu_charging_station_backend.entity.ReportInfo;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,11 +18,11 @@ public class ChargingStationResponse {
     private String name;
     private String description;                  // 充电桩描述
     private Integer mode;
-    private Double power;
-    private Double serviceFee;                   // 服务费单价
-    private Double peakPrice;                    // 峰时，10:00~15:00, 18:00~21:00
-    private Double normalPrice;                  // 平时，7:00~10:00, 15:00~18:00, 21:00~23:00
-    private Double valleyPrice;                  // 谷时，23:00~7:00
+    private BigDecimal power;
+    private BigDecimal serviceFee;                   // 服务费单价
+    private BigDecimal peakPrice;                    // 峰时，10:00~15:00, 18:00~21:00
+    private BigDecimal normalPrice;                  // 平时，7:00~10:00, 15:00~18:00, 21:00~23:00
+    private BigDecimal valleyPrice;                  // 谷时，23:00~7:00
     private Integer maxQueueLength;              // 最大排队数
     private ChargingStationSlot slot;            // 包含 实时状态+队列
     private ReportInfo report;                   // 报表数据
